apps/api/pyproject.toml
[project]
name = "myetal-api"
version = "0.1.0"
description = "MyEtal backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pyjwt>=2.10.0",
    "argon2-cffi>=23.1.0",
    "email-validator>=2.2.0",
    "qrcode[pil]>=7.4.0",
    "httpx>=0.27.0",
    "sentry-sdk[fastapi]>=2.18.0",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
    "cachetools>=5.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myetal_api"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S"]
# N818: exception names without "Error" suffix (we follow the auth-lib idiom:
#   `InvalidCredentials`, `EmailAlreadyRegistered`, etc.)
# N806: PascalCase locals (we use `SessionLocal` per SQLAlchemy convention)
# S101: `assert` allowed (production code never asserts; tests legitimately do)
# S105/S106: hardcoded-password heuristic flags every string literal "password",
#   "token_type", "token_url" kwargs, and our intentionally-public
#   DEV_SECRET_PLACEHOLDER. Noise.
# B008: function call in argument default. FastAPI's idiomatic
#   `Query(...)`, `Depends(...)`, `Body(...)` usage in defaults is the
#   blessed pattern; ruff treats them like mutable-default-arg footguns.
ignore = ["S101", "S105", "S106", "N818", "N806", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
